Refactor: separate guard module
this is for later reuse in rm
