+++ trashcli/empty/empty_cmd.py
@@ -7,7 +7,7 @@ from trashcli.empty.console import Console
 from trashcli.empty.empty_action import EmptyAction, EmptyActionArgs
 from trashcli.empty.errors import Errors
 from trashcli.empty.existing_file_remover import ExistingFileRemover
-from trashcli.empty.is_input_interactive import is_input_interactive
+from trashcli.guard.is_input_interactive import is_input_interactive
 from trashcli.empty.parser import Parser
 from trashcli.empty.print_time_action import PrintTimeAction, PrintTimeArgs
 from trashcli.fs import ContentsOf
