+++ trashcli/guard/is_input_interactive.py
